+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Moving `key` function into `IFormula`
 - **Breaking**: Remove `state.noEffects()` extension function. Use `transition(state)` instead.
 - Added `StreamFormula` and `ObservableFormula`.
+- Allow nullable formula state.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.